requirements.txt
python =3.12.1
Django~=5.0.8
django-browser-reload~=1.13.0
sqlparse~=0.5.1
asgiref~=3.8.1
whitenoise
cryptography-43.0.3
